+++ bauh/app.py
@@ -81,3 +81,7 @@ def main():
     cache_cleaner.start()
 
     sys.exit(app.exec_())
+
+
+if __name__ == '__main__':
+    main()